feat&refactor: AI 분석 결과 페이지 라우팅 및 AI 분석 라우팅 방식 변경
