+++ web/src/pages/analyze/AnalyzeResultPage.tsx
@@ -1,0 +1,3 @@
+export const AnalyzeResultPage = () => {
+  return <div>123</div>;
+};
