+++ web/src/pages/index.ts
@@ -5,8 +5,9 @@ export { CommunityWrite } from "./communityWrite/CommunityWrite";
 export { LoginPage } from "./auth/LoginPage";
 export { SignUpPage } from "./auth/SignUpPage";
 export { SearchPage } from "./searchPage/SearchPage";
-export { AnalyzeLoadingPage } from "./loading/AnalyzeLoadingPage";
+export { AnalyzeLoadingPage } from "./analyze/AnalyzeLoadingPage";
 export { AnalyzePage } from "./analyze/AnalyzePage";
+export { AnalyzeResultPage } from "./analyze/AnalyzeResultPage";
 export { ContentListPage } from "./content/ContentListPage";
 export { ArticleDetailPage } from "./content/ArticleDetailPage";
 export { InterviewDetailPage } from "./content/InterviewDetatilPage";
