+++ web/src/routes/AppRouter.tsx
@@ -13,6 +13,7 @@ import {
   ContentListPage,
   ArticleDetailPage,
   InterviewDetailPage,
+  AnalyzeResultPage,
 } from "@/pages";
 
 import { BottomLayout } from "@/layout/BottomLayout";
@@ -54,14 +55,6 @@ const AppRouter = createBrowserRouter([
     ],
   },
   {
-    path: path.loading,
-    element: (
-      <Layout>
-        <AnalyzeLoadingPage />
-      </Layout>
-    ),
-  },
-  {
     //TODO: @tifsy 커뮤니티 글 상세보기 path 변경
     path: path.community.detail(":id"),
     element: <CommunityDetail />,
@@ -96,11 +89,21 @@ const AppRouter = createBrowserRouter([
   },
   {
     path: path.analyze.base,
-    element: (
-      <Layout>
-        <AnalyzePage />
-      </Layout>
-    ),
+    element: <Layout />,
+    children: [
+      {
+        path: "",
+        element: <AnalyzePage />,
+      },
+      {
+        path: path.analyze.specific.loading,
+        element: <AnalyzeLoadingPage />,
+      },
+      {
+        path: path.analyze.specific.result(":analyzeId"),
+        element: <AnalyzeResultPage />,
+      },
+    ],
   },
 ]);
